Add basic indexing support
Non-ground references are lazily indexed during evaluation if they are
contained in an equality expression where the other side is ground (or ground
after plugging the current bindings).

Updates to data that is indexed cause the built index to be dropped from the
indices.

The indexing supports exact match, reverse lookup using the ground value in
the expression. The result of the lookup is a set of bindings that when
plugged into the reference, result in the equality evaluating to true. In
this way, the indexing is not "lossy", i.e., it does not produce false matches
(positive or negative).