Fix safety check for aliased imports
Reorder the local safety checks to occur before references have been resolved.
The local safety checks can happen before the refrences are resolved because
they do not depend on the values of fully qualified names (unlike, e.g., the
recursion check).

In this case, the safety check on the rule body needs ot happn before
references are resolved because the head of the fully qualified may not be in
the compiler's Globals map.

- Add note about built-in operator check
- Refactor compiler stages slightly so that tests are more maintainable