Fix global variable support in topdown evaluation
Previously, global variable bindings were not propagated to child contexts
during evaluation. For instance:

import req
p :- q
q :- req = true

>> eval(p, {req: true})

In this case, the binding for req would not be available when evaluating the
body of q. As a result, the query would fail.

Now, we maintain two separate sets of bindings: global and local. The global
bindings are automatically propagated to the child context. Locations that
would previously use ctx.Bindings.Get() have been updated to use ctx.Binding()
which consults both the global and local bindings.