Add support for vdoc references without suffix
When the reference refers to a virtual document without accessing an
element/attribute of the document, the rule must be fully evaluated and
the result must be bound in the current context.

For example, given two packages:

package a.b.c
p[x] = y :- ... // something that defines x's and y's

package x.y.z
import data.a.b.c.p
q = true :- p = x, // do something with x

In the second package, the reference to "p" resolves to "data.a.b.c.p". In the
context of evaluating references, we say this rule reference has no suffix.

Before, this would result in an error at evaluation time. Now, topdown
evaluates "p" fully and binds the result in the current context.