Fix recompile bug
After refs have been resolved, aliases need to be removed otherwise we get
false positive (unsafe variable) errors. It's intended that the policies kept
in the PolicyStore can be recompiled. Since the PolicyStore keeps the compiled
version of the policy, we need to make sure compiled policies can be
recompiled. Eventually this should become a noop.