Show term values when possible
Previously, the REPL treated all bodies as queries: the query was evaluated
and for each instance where the query was true, the variable bindings were
displayed. If there were no variables in the body, "true" was printed. If
there were no instances where the query was true, "false" was printed.

Now, the REPL treats single term expressions as a special case under the
assumption that the user wants to see the value of the term. In cases where
the term can only evaluate to a single value (e.g., it's ground), the value of
the term is serialized to JSON and printed. This is useful for initially
learning the language and eventually experimenting with references, e.g.:

> data.servers[0].name
"web"
>

In cases where the term can evaluate to multiple values (i.e., it's
non-ground), the REPL displays one row of output for each instance where the
query is true. Each row contains the bindings of non-wildcard variables as
well as a column for the value of the term itself. This is useful for quickly
evaluating references without defining an intermediate variable, e.g.:

> data.servers[x].name
0,"web"
1,"app"
...
>