Fix: Dereferencing refs embedded in collections
Previously, the code that deals with references to virtual docs was not
handling cases where virtual docs were produced with references embedded
inside them. The workaround for this was to assign the reference element
embedded in the collection to an intermediate variable and then dereference
that variable. E.g.,

q[k] = vs :- k = "...", vs = [ r | r = data.a[_] ]

p :- q["..."][i].a.b.c > 1

In this case, the "vs" array will contain references to elements in data.a.
Assuming those elements are themselves objects, the reference in "p" will
iterate over the elements and dereference them with "[i].a.b.c".

The old code using the ast.Query function would not handle this correctly: the
ast.Query function would encounter the reference elements and treat the query
as undefined (thus "p" would be undefined).

The workaround assigned the collection elements to an intermediate var, e.g.:

p :- q["..."][i] = x, x.a.b.c > 1

By assignign the collection elements to an intermediate var, the ast.Query
function would not see the reference elements and the query would work.

Now, the code in topdown handles these embedded references correctly by
recursively processing the virtual docs and evaluating any embedded
references.