Update topdown to handle composites in rule head
This change provides a 10x speedup in the scheduler benchmark because each
selector_match query can avoid scanning all pods and rcs (because the pod_id
and rc_id values are provided as inputs in the normal case.)

Set rules are handled by unifying the value in the head with the value of the
key in the current context.

Object rules are handled in the same way as before.

Rewrote a few comments to capture future improvements to unification and
improvements needed in the topdown.Query implementation.