Handle duplicate object keys in topdown.Query
When an object document is queried, if the rule generates multiple values for
the same key, it is supposed to be treated as an error. This is already done
for object document references w/o a suffix. The only remaining place is when
an object document is being iterated as part of a query.