Add caching to topdown evaluation
The cache currently contains virtual documents generated by rules producing
partial object and complete docs.

The cache is consulted before evaluating the rule as part of reference
evaluation. If a cache miss occurs, the rule is evaluated as normal and then
the result is put into the cache. If a cache hit occurs, the cached value is
used and the rule is not re-evaluated.

This provides another ~10x speedup in the scheduler benchmark.