Refactor performance tests a bit
The scheduler benchmark is useful as an end-to-end test of OPA so it is nice
to have it run as part of the normal suite (as long as it doesn't take too
long). To accomodate this, these changes split the scheduler benchmark into
two suites: (1) normal test that runs on captured dataset (10 nodes x 30
pods) and (2) benchmark test that pre-populates storage with nodes, pods, RCs,
etc. This allows us to simulate larger data sets (e.g., 1k nodes x 3k pods) in
the benchmark.