Improve query evaluation
These changes improve query evaluation in two main ways.

The first is that the storage layer is no longer accessed at each step in the
evaluation of a single reference. Before, topdown would issue a read for
each element in the reference, starting from the head, e.g., for data.a.b.c,
topdown would execute four reads. This was problematic for mounted stores. The
new implementation is greedy, meaning it executes reads on the longest ground
portion of the reference it has available. This will avoid the unnecessary reads
against mounted stores.

The second is that the implementation now handles references that iterate over
virtual documents or implicitly refer to base and virtual documents. This
will allow for users to execute queries against the root of the global
document and see the entire extent OPA's data set. This will probably be quite
expensive but for development purposes, it will likely be useful.

Lastly, the topdown.Query interface has been simplified. There is still a
special case for querying sets, but that will be removed in time.