Fix bug in ad-hoc query execution over HTTP
The server was using the cached compiler to compile ad-hoc queries. The
problem with this was that an invalid ad-hoc query would invalidate the
cached compiler (thus breaking subsequent use of it).

Since ad-hoc queries do not have to use the cached compiler, simply don't use
it. If ad-hoc queries need to make use of the cached compiler in the future,
we can change the compiler to not invalidate data structures, be immutable,
etc.