Remove undefined response from API
This change is not backwards compatible.

The server no longer returns an object {"IsUndefined": true} in response to
GET against /data/... that is missing/undefined.

Previously, the server would return this object to indicate to callers that a
virtual document was undefined as opposed to the case where a base document
was missing. However, with the recent changes to support queries that
implicitly refer to base and virtual documents (e.g., GET /data) the topdown
implementation no longer returns a storage error in case of a missing base
document. As a result, there is no need for the {"IsUndefined": true} response
anymore.

Also, fixing a couple small mistakes in the REST API docs.