Fix compiler error reporting
Before, callers would either flatten errors to a string or pick the first
error reported by the compiler. This was a bad approach and lead to an issue
for users running OPA from the command line, for example:

 $ opa run test.rego

If test.rego failed to compile, the run command would only print the first
error...

This change lets callers treat all compiler errors as a single error and
factors the flattening into the Error() implementation of the new ast.Errors
type.