Add explanations to REST API
GET /v1/data[/<path>]?explain=full modifies the query to return the full trace
from the query evaluation.

GET /v1/data[/<path>]?explain=truth modifies the query to return the
truth-explanation for the query. If the path is undefined, the status code
will be 404 and the body will be empty.

The trace responses are modelled as an array of objects. The array elements
contain the same fields as the topdown.Event but are defined separately to
avoid API compatibility issues.