Update compiler to support references in head
These changes update the compiler to support references in the head of rules.
Specifically, when a reference is found in the head (either in the key or
value), the associated term is replaced with a variable that will be bound in
the body. This stage is called rewriteRefsInHead in the code.

This new stage is dependant on all references being resolved beforehand. Since
this stage modifies the body, it should be performed before the safety check
on the body is run. This ensures that cases such as the following are handled:

p[data.a[i]] :- i < 100

The example would be initially rewritten as:

p[__local0__] :- i < 100, __local0__ = data.a[i]

The safety check on the body will handle the reordering that is required so
that the expression "i < 100" is safe.

Lastly, because references are resolved before the safety checks run, the
Globals field is no longer needed--it was only used for the safety check up to
now, and it's probably better to keep the API surface area as small as
possible. By association the Exports field can be removed as well. That being
said, the logic required to compute the globals for each module is still
needed for resolving references at the beginning.