Fix evaluation of refs to set literals
This change addresses two issues with evaluation of refs to set literals:

1) If the set or lookup value contained references, they were not being
resolved before checking for membership. So for example,
{data.a[0]}.contains("a") would fail even if data.a[0] = "a". This is solved
by simply resolving all references in the set and lookup value before
performing the membership check.

2) Set literals embedded inside other collections were not being evaluated.
For example, [{"a", "b"}] = x, x[0]["a"] would be false. This solved by
refactoring the branch that handles sets to behave more closely to the ones
for arrays and objects.

Fixes #149