Fix parser to transform data var into ref
Previously, the token "data" would be parsed into Var("data"). Because "data"
is a reserved word, it would pass the safety checks and then fail to evaluate.

With this change, Var("data") is transformed into Ref("data"). This means the
safety checks will catch the recursion (instead of failing during evaluation).