Fix ast.Number hash and equality
The equality check was broken for exponent floats when the underlying
representation was changed to use json.Number. The new code uses the Compare
function which is correct. There doesn't appear to be a big impact on
performance at this point.

The hash function would panic on numbers bigger than 64-bit float. Fallback to
string hash in this case. This is somewhat unlikely in practice but it must be
handled none the less.