Add POST operation to Data API
This allows callers to specify the input document in the message body
which is more robust and less error prone than using GET with query
parameters. The latter requires URL encoding, unfamiliar cURL flags,
etc.

We will keep the input parameter in the GET API for convenience but
documentation and examples will prefer the POST API.