Add high level API for Go users
Package rego is a high-level API that should make it easier for Go users
to integrate with OPA. The package does not require that users
instantiate storage, compiler, etc. Users can pass in queries and
modules as strings and obtain results as native Go values.

The package is designed so that different aspects of evaluation can be
customized easily.