Refactor to always build and use the rule indices
This change modifies the compiler to always build indices for rules
(regardless of rule set size) and topdown to always perform an index
lookup.

This will make the else keyword changes in topdown easier.

Also, fix bug in evalRefRecTree that had evalRefRule called with an
incorrect path. Since evalRefRule is always called with the same
ref/path pair from evalRefRecTree, this did not result in incorrect
results being produced.