Fix eval bug introduced by rule indexing
Queries to evaluate an entire partial set or object should return an
empty value as opposed to being undefined. In the event that the rule
index lookup finds no matching rules, the caller (topdown) must know
what type of rules *could* be returned so it can decide how to proceed.