Correct data race in the parser extensions
When the parser is post-processing parsed Rego, it takes steps to
transform variable references to the `data` and `input` root documents
into Refs with the variables as the head. It does this using the
Transform function of the ast package, which will proceed to transform
the new Ref as well. While it would always replace the ref head with
itself, the Go race detector correctly saw concurrent reads and writes
to the Ref as a data race. The parser extensions have been corrected to
replace variable instances of the root documents with separate copies of
their Ref counterparts, instead of using a single global instance of
each across calls.