Simplify input query parameters in data GET requests
The data GET endpoint supported multiple forms to supply an input
document with, but this was too complicated. It's much simpler to simply
require the entire input document to be defined at once, and so the
parameter has been changed to only support the definition of the entire
input document, like `...?input={"foo":"bar"}`.