Fix image version in HTTP API tutorial
Before we were using the "latest" tag in the tutorial. This caused an
issue because we thought the JWT token support added to the toy server
had been built and published under the latest tag however it had not.

Since the toy server functionality is somewhat coupled to the tutorial,
hardcoding the version number here is acceptable.